update Course model and add migrations

+++ courses/models.py
@@ -12,7 +12,8 @@ class Course(models.Model):
     description = models.TextField(verbose_name='описание', **NULLABLE)
     owner = models.ForeignKey(User, on_delete=models.CASCADE,
                               null=True, verbose_name='пользователь')
-    price = models.FloatField(verbose_name='стоимость', **NULLABLE)
+    price = models.DecimalField(max_digits=6, decimal_places=2,
+                                verbose_name='стоимость', **NULLABLE)
 
     def __str__(self):
         return self.name
